初步完成 User 相关服务的编写

+++ pkg/errno/errno.go
@@ -15,7 +15,7 @@ const (
 )
 
 type ErrNo struct {
-	ErrCode int64
+	ErrCode int32
 	ErrMsg  string
 }
 
@@ -23,7 +23,7 @@ func (e ErrNo) Error() string {
 	return fmt.Sprintf("err_code=%d, err_msg=%s", e.ErrCode, e.ErrMsg)
 }
 
-func NewErrNo(code int64, msg string) ErrNo {
+func NewErrNo(code int32, msg string) ErrNo {
 	return ErrNo{code, msg}
 }
